Remove FolderInput
