+++ packages/ndla-ui/src/MyNdla/index.ts
@@ -7,5 +7,4 @@
  */
 
 import Folder from "./Resource/Folder";
-import FolderInput from "./Resource/FolderInput";
-export { Folder, FolderInput };
+export { Folder };
